define independent context invalidation service, move configuration to xml
