+++ src/Resources/config/user_context.xml
@@ -25,10 +25,18 @@
             <argument type="service" id="security.token_storage" on-invalid="ignore" />
         </service>
 
-        <service id="fos_http_cache.user_context.logout_handler" class="FOS\HttpCacheBundle\Security\Http\Logout\ContextInvalidationLogoutHandler">
+        <service id="fos_http_cache.user_context_invalidator" class="FOS\HttpCacheBundle\UserContextInvalidator">
             <argument type="service" id="fos_http_cache.default_proxy_client" />
             <argument />
             <argument />
+        </service>
+
+        <service id="fos_http_cache.user_context.logout_handler" class="FOS\HttpCacheBundle\Security\Http\Logout\ContextInvalidationLogoutHandler" public="false">
+            <argument type="service" id="fos_http_cache.user_context_invalidator" />
+        </service>
+
+        <service id="fos_http_cache.user_context.session_logout_handler" class="FOS\HttpCacheBundle\Security\Http\Logout\ContextInvalidationSessionLogoutHandler" public="false">
+            <argument type="service" id="fos_http_cache.user_context_invalidator" />
         </service>
 
         <service id="fos_http_cache.user_context.anonymous_request_matcher" class="FOS\HttpCacheBundle\UserContext\AnonymousRequestMatcher">
